% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_multinomial.R
\name{agg_multinomial}
\alias{agg_multinomial}
\title{Aggregate data for the model using a multinomial distribution}
\usage{
agg_multinomial(
  data,
  responses,
  group = NULL,
  DV_name = "y",
  nDV_name = "nRet",
  numeric = TRUE
)
}
\arguments{
\item{data}{A data frame containing the data to be aggregated}

\item{responses}{A character vector containing the response variables}

\item{group}{The grouping variable}

\item{DV_name}{The name of the dependent variable, if NULL, the response variables will not be combined.}

\item{nDV_name}{The name of the number of retrieval}

\item{numeric}{A logical value indicating whether to rename the columns of the dependent variable}
}
\value{
A data frame containing the aggregated data
}
\description{
This function aggregates data for the model using a multinomial distribution.
}
