% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/append_job.R
\name{append_job}
\alias{append_job}
\title{Append a new job to the job log}
\usage{
append_job(name = NULL, useCore = 1, untilFinished = FALSE, priority = 0)
}
\arguments{
\item{name}{The name of the job.}

\item{useCore}{The number of cores used by the job.}

\item{untilFinished}{Boolean value indicating whether the job should not run
until previous jobs are finished.}

\item{priority}{The priority of the job.}

\item{status}{The status of the job.}

\item{startTime}{When the job starts.}

\item{endTime}{When the job ends.}

\item{duration}{The duration of the job.}
}
\description{
This function will append a new job to the job log.
}
