% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_class.R
\docType{data}
\name{BySpeLFD}
\alias{BySpeLFD}
\title{BySpeLFD}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
BySpeLFD
}
\value{
This function returns the 'sampleLFDbyspe' object.
}
\description{
The \code{BySpeLFD} class implements the class of SMART to handle species samplings.
}
\keyword{datasets}

