% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_each.R
\name{compare_each}
\alias{compare_each}
\title{Run the model comparison and calculate Bayes Factor}
\usage{
compare_each(pars, sample_path, sample_prefix, favorBF = 3)
}
\arguments{
\item{pars}{A list of parameters to compare. Each parameter is a list of values.}

\item{sample_path}{The path to store the sample.}

\item{sample_prefix}{The prefix of the sample.}

\item{favorBF}{The favor of the Bayes Factor. The default is `3`.}
}
\value{
A table of the model comparison.
}
\description{
This function does all the model comparisons and counts the wins, fails, and unclear times for each model.
}
