% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forms.R
\name{create_gsheet_form}
\alias{create_gsheet_form}
\title{Create a Google Sheets data entry form}
\usage{
create_gsheet_form(name, sheets = NULL, view = TRUE)
}
\arguments{
\item{name}{A string to name the newly created spreadsheet.}

\item{sheets}{Optional input for initialising worksheets.}

\item{view}{Logical. Should the newly created sheet be opened in the
current browser? Default to TRUE.}
}
\value{
A spreadsheet with specified name and specified sheets
  in user's Google Sheets account.
}
\description{
Create a Google Sheets data entry form
}
\examples{
## Create a Google Sheet named "village_list"
if (googlesheets4::gs4_has_token()) {
  create_gsheet_form("SMART Survey")
}

}
\author{
Ernest Guevarra
}
