% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jatos-helper.R
\name{download_all_data}
\alias{download_all_data}
\title{Download all data for specified batch IDs}
\usage{
download_all_data(data_url, headers, batch_ids, data_path)
}
\arguments{
\item{data_url}{URL for downloading data}

\item{headers}{HTTP headers including authorization}

\item{batch_ids}{Vector of batch IDs}

\item{data_path}{Path where data will be saved}
}
\description{
Helper function to download all data for the specified batch IDs.
}
\keyword{internal}
