% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_JATOS_data.R
\name{download_metadata}
\alias{download_metadata}
\title{Download metadata from JATOS}
\usage{
download_metadata(url, study_id, headers, metadata_path)
}
\arguments{
\item{url}{Base URL for the JATOS API}

\item{study_id}{ID of the study}

\item{headers}{HTTP headers including authorization}

\item{metadata_path}{Path where metadata will be saved}
}
\value{
A data frame containing the metadata
}
\description{
Helper function to download metadata from the JATOS server.
}
\keyword{internal}
