% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/facet_scales.R
\name{facet_scales}
\alias{facet_scales}
\title{Custom facet scales}
\usage{
facet_scales(xlims = NULL, ylims = NULL)
}
\arguments{
\item{xlims}{A list of numeric vectors specifying the x-axis limits for each facet.
Each element should be a vector of length 2 (c(min, max)) or NULL to use default limits.}

\item{ylims}{A list of numeric vectors specifying the y-axis limits for each facet.
Each element should be a vector of length 2 (c(min, max)) or NULL to use default limits.}
}
\value{
A ggplot2 layer that can be added to a ggplot object
}
\description{
Sets different axis limits for each facet in a ggplot2 facet_wrap or facet_grid plot.
}
\details{
This is an advanced function that modifies the internal behavior of ggplot2.
It should be used with caution as it may break with future ggplot2 updates.
The function works by intercepting the initialization of scales in the facet
and replacing the limits with custom values provided by the user.

This function is based on the solution provided at:
https://stackoverflow.com/questions/51735481/ggplot2-change-axis-limits-for-each-individual-facet-panel
}
\examples{
\dontrun{
library(ggplot2)

# Create a sample plot with facets
data <- data.frame(
  x = 1:100,
  y = rnorm(100),
  group = rep(letters[1:4], each = 25)
)

# Set custom y-axis limits for each facet
ylims <- list(c(-1, 1), c(-2, 2), c(-3, 3), c(-4, 4))

# Apply custom limits using the + operator
p <- ggplot(data, aes(x, y)) +
  geom_point() +
  facet_wrap(~group) +
  facet_scales(ylims = ylims)

# Set both x and y limits
xlims <- list(c(1, 20), c(5, 25), c(10, 30), c(15, 35))
p2 <- ggplot(data, aes(x, y)) +
  geom_point() +
  facet_wrap(~group) +
  facet_scales(xlims = xlims, ylims = ylims)
}
}
\author{
Canadian Marine, Chenyu Li
}
