% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_class.R
\docType{data}
\name{FisheryBySpecie}
\alias{FisheryBySpecie}
\title{FisheryBySpecie}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FisheryBySpecie
}
\value{
This function returns the 'FisheryBySpecie' object.
}
\description{
The \code{FisheryBySpecie} class implements the class of SMART to
 handle species samplings.
}
\keyword{datasets}

