% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_class.R
\docType{data}
\name{FishFleet}
\alias{FishFleet}
\title{FishFleet}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
FishFleet
}
\value{
This function returns the 'fishFleet' object.
}
\description{
The \code{FishFleet} class implements the class of SMART to manage fleet data.
}
\keyword{datasets}

