% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ggplot-helper.R
\name{geom_errorbar_adjusted}
\alias{geom_errorbar_adjusted}
\title{Geom errorbar with adjusted width}
\usage{
geom_errorbar_adjusted(
  mapping = NULL,
  data = NULL,
  group_var = NULL,
  width = 0.05,
  ...
)
}
\arguments{
\item{mapping}{Aesthetic mapping}

\item{data}{A data frame}

\item{group_var}{Name of the grouping variable (unquoted)}

\item{width}{Base width value to be multiplied by group size}

\item{...}{Other arguments passed to geom_errorbar}
}
\value{
A ggplot2 layer
}
\description{
Create error bars with width adjusted based on group sizes
}
