% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_JATOS_data.R
\name{get_JATOS_data}
\alias{get_JATOS_data}
\title{Download data from JATOS}
\usage{
get_JATOS_data(
  token,
  url = "https://coglab.xyz/jatos/api/v1/results",
  studyId,
  batchId,
  dataPath = NULL,
  extractInfo = NULL,
  attachments = FALSE
)
}
\arguments{
\item{token}{An API token, you can create a new one on JATOS. Details see: https://www.jatos.org/JATOS-API.html#personal-access-tokens}

\item{url}{Server address. The default value is the address of our lab server.}

\item{studyId}{A unique code of study, you can find it on JATOS}

\item{batchId}{A unique code of a batch session in the study.}

\item{dataPath}{A path used to save data. If NUll, data will be saved in working directory.}

\item{extractInfo}{A list of keys to extract from the data file. Default is NULL.}
}
\value{
data.frame
}
\description{
This function downloads result data from a JATOS server using an API token and specified study ID and batch ID.
It then unzips and extracts the relevant data, including metadata, file names, and file contents.
The function returns a data frame with the relevant data.
}
