% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{get_pps}
\alias{get_pps}
\title{Perform proportional to population size (PPS) sampling based on SMART
approach}
\usage{
get_pps(df, pop, m)
}
\arguments{
\item{df}{A data.frame of clusters/villages with their population size}

\item{pop}{A character value specifying the name of the variable in `df`
for the population size information}

\item{m}{Number of clusters}
}
\value{
A data.frame which is a subset of `df` of the selected clusters for
  the survey sample. A new variable called `cluster_id` is created which
  identifies the cluster number and the reserved clusters.
}
\description{
Perform proportional to population size (PPS) sampling based on SMART
approach
}
\examples{
get_pps(df = clusterList, pop = "pop", m = 30)

}
