% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{get_si}
\alias{get_si}
\title{Get sampling interval}
\usage{
get_si(pop, su)
}
\arguments{
\item{pop}{Total population}

\item{su}{Number of sampling units to be selected}
}
\value{
An integer value for the sampling interval
}
\description{
Get sampling interval
}
\examples{
get_si(villageList1$pop, su = 30)

}
