% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampsize.R
\name{get_ss_hh}
\alias{get_ss_hh}
\title{Calculate number of households to sample given a target sample size of
children 6-59 months old}
\usage{
get_ss_hh(ss, hh_size = 5, u5 = 0.15, f = 0.9)
}
\arguments{
\item{ss}{Target sample size of children 6-59 months old}

\item{hh_size}{Average household size. Default to 5 household members}

\item{u5}{Proportion of the population who are under 5 years old. Default to
0.15 (15\%)}

\item{f}{Proportion of under 5 who are 6-59 months in age. Default to 0.9
(90\%)}
}
\value{
A numeric value for number of households to sample
}
\description{
Calculate number of households to sample given a target sample size of
children 6-59 months old
}
\examples{
get_ss_hh(ss = 500, hh_size = 6.2)

}
