% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/init_job.R
\name{init_job}
\alias{init_job}
\title{Initialize job log.}
\usage{
init_job(path = NULL)
}
\arguments{
\item{path}{The path to the job log.}
}
\description{
This function will create a data frame including model information and store it as a temporary file.
}
