% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_manager.R
\name{init_job}
\alias{init_job}
\alias{view_job}
\alias{remove_job}
\title{Job Log Management}
\usage{
init_job(path = NULL)

view_job(.x = NULL, path = NULL)

remove_job(.x, path = NULL)
}
\arguments{
\item{path}{Character. Optional. The directory where the job log is stored. If `NULL`, the function looks in the temporary directory.}

\item{.x}{Character or numeric. The job name (character) or index (numeric). For `view_job` and `remove_job`, if `NULL`, returns or removes all jobs.}
}
\description{
These functions manage a job log, which is stored as an RDS file. The job log tracks the status, priority, and other details of jobs.
}
\details{
- `init_job`: Initializes a new job log.
- `view_job`: Retrieves the status of one or more jobs.
- `remove_job`: Removes one or more jobs from the job log.
}
\examples{
\dontrun{
# Initialize a job log
init_job()  # Saves in temp directory

# You have to use `smart_runFun` to add jobs to the log

# View job status
view_job()  # View all jobs
view_job(1)  # View job with index 1
view_job("job_name")  # View job with name "job_name"

# Remove a job
remove_job(1)  # Remove job with index 1
remove_job("Task_A")  # Remove job with name "Task_A"
}
}
