% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mutate_form.R
\name{mutate_form}
\alias{mutate_form}
\title{add new columns to data frame using formulas}
\usage{
mutate_form(.data, formulas)
}
\arguments{
\item{.data}{A data frame or data frame extension (e.g. a tibble).}

\item{formulas}{A formula or a vector of formulas to be used for calculating new columns.}
}
\value{
A data frame with new columns added.
}
\description{
This function is used to add new columns to a data frame using formulas.
}
