% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesfactor-helper.R
\name{print.pairwise_comparison}
\alias{print.pairwise_comparison}
\title{Print method for pairwise_comparison objects}
\usage{
\method{print}{pairwise_comparison}(x, ...)
}
\arguments{
\item{x}{A pairwise_comparison object}

\item{...}{Additional arguments passed to print methods}
}
\value{
x invisibly
}
\description{
Print method for pairwise_comparison objects
}
