% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/bayesfactor-helper.R
\name{print.pairwise_comparison_list}
\alias{print.pairwise_comparison_list}
\title{Print method for pairwise_comparison_list objects}
\usage{
\method{print}{pairwise_comparison_list}(x, ...)
}
\arguments{
\item{x}{A pairwise_comparison_list object}

\item{...}{Additional arguments passed to print methods}
}
\value{
x invisibly
}
\description{
Print method for pairwise_comparison_list objects
}
