% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_json_data.R
\name{read_json_data}
\alias{read_json_data}
\title{Read json data and transform it into a data frame}
\usage{
read_json_data(files)
}
\arguments{
\item{files}{A vector of json files to read}
}
\value{
a data frame combined all the data
}
\description{
This function reads a list of json files and combines them into a single data frame.
}
