% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_metadata.R
\name{read_metaData}
\alias{read_metaData}
\title{Read the files information from the metadata file}
\usage{
read_metaData(path)
}
\arguments{
\item{path}{The path to the directory containing the metadata file}
}
\value{
data.frame
}
\description{
This function reads the metadata file and extracts the relevant information
}
