% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_manager.R
\name{remove_job}
\alias{remove_job}
\title{Remove job(s) from the job log.}
\usage{
remove_job(.x, path = NULL)
}
\arguments{
\item{.x}{Character or numeric. The job name (character) or index (numeric) of the job(s) to be removed.}

\item{path}{Character. Optional. The directory where the job log is stored.
If `NULL`, the function looks in the temporary directory.}
}
\description{
Removes a job (or multiple jobs) from the job log by index or name.
}
\examples{
remove_job(1)  # Remove job with index 1
remove_job("Task_A")  # Remove job with name "Task_A"

}
