% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{roll_sum_range}
\alias{roll_sum_range}
\title{Get the cumulative population size range}
\usage{
roll_sum_range(v, na = c("na.rm", "na.omit"), as_char = FALSE)
}
\arguments{
\item{v}{A vector of population sizes to apply the cumulative population
size range calculations on}

\item{na}{How should missing values be treated? Either `na.rm` to remove or
`na.omit` to omit by assigning a value of 0. Default is to omit.}

\item{as_char}{Logical. Should output be a character vector or a
data.frame of two values. Default to FALSE.}
}
\value{
A vector of cumulative population size range
}
\description{
Get the cumulative population size range
}
\examples{
roll_sum_range(v = clusterList$pop)

}
