% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_class.R
\docType{class}
\name{SampleMap}
\alias{SampleMap}
\title{SampleMap}
\format{\code{\link{R6Class}} object.}
\usage{
SampleMap
}
\value{
Object of \code{\link{R6Class}} with attributes and methods for the Environmental data.
}
\description{
The \code{SampleMap} class implements the class of SMART
to control geographical data.
}
\section{Fields}{

\describe{
\item{\code{gridPath}}{Missing description.}

\item{\code{gridName}}{Missing description.}

\item{\code{gridShp}}{Missing description.}

\item{\code{gridBbox}}{Missing description.}

\item{\code{gridBboxExt}}{Missing description.}

\item{\code{gridBboxSP}}{Missing description.}

\item{\code{areaGrid}}{Missing description.}

\item{\code{areaStrata}}{Missing description.}

\item{\code{weightStrata}}{Missing description.}

\item{\code{harbDbf}}{Missing description.}

\item{\code{bioPath}}{Missing description.}

\item{\code{bioName}}{Missing description.}

\item{\code{bioShp}}{Missing description.}

\item{\code{bioDF}}{Missing description.}

\item{\code{gridPolySet}}{Missing description.}

\item{\code{gridFortify}}{Missing description.}

\item{\code{nCells}}{Missing description.}

\item{\code{sCells}}{Missing description.}

\item{\code{griCent}}{Missing description.}

\item{\code{gridBathy}}{Missing description.}

\item{\code{centDept}}{Missing description.}

\item{\code{clusInpu}}{Missing description.}

\item{\code{clusMat}}{Missing description.}

\item{\code{indSil}}{Missing description.}

\item{\code{indCH}}{Missing description.}

\item{\code{cutFG}}{Missing description.}

\item{\code{tmpCut}}{Missing description.}

\item{\code{availData}}{Missing description.}

\item{\code{rawInpu}}{Missing description.}

\item{\code{cutResult}}{Missing description.}

\item{\code{cutResEffo}}{Missing description.}

\item{\code{cutResShp}}{Missing description.}

\item{\code{cutResShpCent}}{Missing description.}

\item{\code{cutResShpFort}}{Missing description.}

\item{\code{fgWeigDist}}{Missing description.}

\item{\code{ggBioDF}}{Missing description.}

\item{\code{ggDepth}}{Missing description.}

\item{\code{ggDepthFGbox}}{Missing description.}

\item{\code{ggEffoFGbox}}{Missing description.}

\item{\code{ggEffoFGmap}}{Missing description.}

\item{\code{ggBioFGmat}}{Missing description.}

\item{\code{ggCutFGmap}}{Missing description.}

\item{\code{ggIchFGlin}}{Missing description.}

\item{\code{ggSilFGlin}}{Missing description.}

\item{\code{ggBetaFGmap}}{Missing description.}

\item{\code{ggBetaFGbox}}{Missing description.}

\item{\code{ggProdFGmap}}{Missing description.}

\item{\code{ggProdFGbox}}{Missing description.}

\item{\code{ggMapFgFishery}}{Missing description.}

\item{\code{ggMapFgSurvey}}{Missing description.}

\item{\code{gooMap}}{Missing description.}

\item{\code{gooMapPlot}}{Missing description.}

\item{\code{gooGrid}}{Missing description.}

\item{\code{gooBbox}}{Missing description.}

\item{\code{sampColScale}}{Missing description.}

\item{\code{plotRange}}{Missing description.}
}}

\section{Methods}{

\describe{
  \item{Documentation}{For full documentation of each method go to https://github.com/smartR_blablabla/}
  \item{\code{initialize(grid_path)}}{This method is used }
  \item{\code{setAreaGrid()}}{This method is used }
  \item{\code{setAreaStrata(vectorStrata)}}{This method is used }
  \item{\code{setWeightStrata()}}{This method is used }
  \item{\code{loadHarbDbf(dbf_path)}}{This method is used }
  \item{\code{set_ggMapFgSurvey(rawSampCoo)}}{This method is used }
  \item{\code{set_ggMapFgFishery(rawSampCoo)}}{This method is used }
  \item{\code{createGridBbox()}}{This method is used }
  \item{\code{getGooMap()}}{This method is used }
  \item{\code{setGooPlot()}}{This method is used }
  \item{\code{setPlotRange()}}{This method is used }
  \item{\code{setGooGrid()}}{This method is used }
  \item{\code{plotGooGrid()}}{This method is used }
  \item{\code{plotGooGridData(grid_data)}}{This method is used }
  \item{\code{setSampColScale(fac_col)}}{This method is used }
  \item{\code{plotGooSpeSur(poi_data)}}{This method is used }
  \item{\code{plotGooSpeFis(poi_data)}}{This method is used }
  \item{\code{setGooBbox()}}{This method is used }
  \item{\code{plotGooBbox()}}{This method is used }
  \item{\code{setGridPath(path2grid)}}{This method is used }
  \item{\code{setGridName()}}{This method is used }
  \item{\code{loadGridShp()}}{This method is used }
  \item{\code{setBioPath()}}{This method is used }
  \item{\code{setBioName()}}{This method is used }
  \item{\code{loadBioShp()}}{This method is used }
  \item{\code{addBioShp(bio_path)}}{This method is used }
  \item{\code{loadBioDF(bio_path)}}{This method is used }
  \item{\code{plotBioDF()}}{This method is used }
  \item{\code{setGgBioDF()}}{This method is used }
  \item{\code{ggplotBioDF()}}{This method is used }
  \item{\code{createPolySet()}}{This method is used }
  \item{\code{fortifyGridShp()}}{This method is used }
  \item{\code{setNumCell()}}{This method is used }
  \item{\code{setGridCenter()}}{This method is used }
  \item{\code{getGridBath()}}{This method is used }
  \item{\code{saveGridBath(bathy_path)}}{This method is used }
  \item{\code{loadGridBath(bathy_path)}}{This method is used }
  \item{\code{getCentDept()}}{This method is used }
  \item{\code{setGgDepth(isoLine)}}{This method is used }
  \item{\code{ggplotGridBathy()}}{This method is used }
  \item{\code{plotSamMap(title, celCol)}}{This method is used }
  \item{\code{plotCoho(abbs)}}{This method is used }
  \item{\code{setClusInpu(whiData, howData)}}{This method is used }
  \item{\code{calcFishGrou(numCuts, minsize, maxsize, modeska, skater_method, nei_queen)}}{This method is used }
  \item{\code{plotFishGrou(ind_clu)}}{This method is used }
  \item{\code{setCutResult(ind_clu)}}{This method is used }
  \item{\code{setDepthFGbox()}}{This method is used }
  \item{\code{setEffoFGbox()}}{This method is used }
  \item{\code{setEffoFGmap()}}{This method is used }
  \item{\code{setBioFGmat()}}{This method is used }
  \item{\code{setCutFGmap()}}{This method is used }
  \item{\code{setIchFGlin(numCut)}}{This method is used }
  \item{\code{setSilFGlin(numCut)}}{This method is used }
  }
}

\keyword{data}
