% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/seq_model_comparsion.R
\name{seq_model_comparsion}
\alias{seq_model_comparsion}
\title{Run a sequence of model comparisons.}
\usage{
seq_model_comparsion(
  fun,
  pars,
  form_fun = NULL,
  model_fun = NULL,
  prior_fun = NULL,
  args,
  favorBF = 3,
  model_name = "Model",
  model_path,
  sample_path,
  bf_path,
  maxCore = NULL,
  sample_check = TRUE
)
}
\arguments{
\item{fun}{The function to run the model supports only the \code{brms} model and the \code{bmm} model.}

\item{pars}{A list of parameters to compare. Each parameter is a list of values.}

\item{form_fun}{The function to generate the formula. The default is \code{NULL}.}

\item{model_fun}{The function to generate the model. This parameter only supports the \code{bmm} model.}

\item{prior_fun}{The function to generate the prior. The default is \code{NULL}.}

\item{args}{The arguments of the function to run in parallel.}

\item{favorBF}{The favor of the Bayes Factor. The default is \code{3}.}

\item{model_name}{The name of the model. The default is \code{Model}.}

\item{model_path}{The path to store the model.}

\item{sample_path}{The path to store the sample.}

\item{bf_path}{The path to store the Bayes Factor.}

\item{maxCore}{The maximum number of cores that can be used to run the function.}

\item{sample_check}{Logical. If \code{TRUE}, the function will check whether the sample exists. The default is \code{TRUE}.}
}
\value{
A table of the model comparison.
}
\description{
This function runs a sequence of model comparisons. It supports the comparison of multiple models with different parameters.
In the same parameter, you can test multiple values. The function will automatically generate all the combinations of the parameters.
the comparsion for the next parameter will be based on the best model of the previous parameter.
}
