% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_gui.R
\name{smart_gui}
\alias{smart_gui}
\title{SMART GUI}
\usage{
smart_gui()
}
\value{
This function does not return a value.
}
\description{
The \code{smart_gui} function implements the main graphical user interface of SMART.
}
