% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_run.R
\name{smart_run}
\alias{smart_run}
\title{Run the code in parallel with the waiting list}
\usage{
smart_run(
  ...,
  untilFinished = FALSE,
  cores = 1,
  maxCore = NULL,
  priority = 1,
  checkInt = 17,
  name = NULL
)
}
\arguments{
\item{...}{The code to run in parallel.}

\item{untilFinished}{Logical. If `TRUE`, the code will not run until the previous code is finished.}

\item{cores}{The number of cores required to run the code.}

\item{maxCore}{The maximum number of cores that can be used to run the code.}

\item{priority}{The priority of the code. The code with a higher priority will be run first.}

\item{checkInt}{The interval to check the job log.}

\item{name}{The name of the job.}
}
\description{
This function is used to run the code in parallel with the waiting list.
It relies on the `job` package to manage the job log.
}
