% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_runCustom.R
\name{smart_runCustom}
\alias{smart_runCustom}
\title{if the model are in the first place
if there are sufficient cores
if `untilFinished` is FALSE, run the model
otherwise, wait for a while and check the job log again}
\usage{
smart_runCustom(
  ...,
  untilFinished = FALSE,
  core = 1,
  maxCore = NULL,
  priority = 1,
  checkInt = 19,
  name = NULL
)
}
\description{
if the model are in the first place
if there are sufficient cores
if `untilFinished` is FALSE, run the model
otherwise, wait for a while and check the job log again
}
