% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_runFun.R
\name{smart_runFun}
\alias{smart_runFun}
\title{Run a Function as a Job with a Waiting List}
\usage{
smart_runFun(
  fun,
  args,
  untilFinished = NULL,
  cores = NULL,
  maxCore = NULL,
  priority = 1,
  checkInt = 17,
  name = NULL,
  export = FALSE
)
}
\arguments{
\item{fun}{A function to be run in parallel.}

\item{args}{A list of arguments for \code{fun}.}

\item{untilFinished}{NULL or a numeric vector. If provided, the function will not run until the specified jobs are completed.}

\item{cores}{The number of cores required to run the function.}

\item{maxCore}{The maximum number of cores available for execution.}

\item{priority}{The priority level of the function. Higher priority jobs are executed first.}

\item{checkInt}{The time interval (in seconds) for checking the job log.}

\item{name}{A name for the job. If \code{NULL}, a random identifier is generated.}

\item{export}{A character vector specifying objects to export to the global environment.}
}
\description{
This function runs a function in parallel while managing execution with a waiting list.
It uses the \code{job} package to manage job logs and execution.
}
