% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_class.R
\docType{data}
\name{SmartProject}
\alias{SmartProject}
\title{SmartProject}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SmartProject
}
\value{
This function returns the 'smartProject' object.
}
\description{
The \code{SmartProject} class implements the main class of
 SMART.
}
\keyword{datasets}

