% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartr.R
\docType{package}
\name{smartr}
\alias{smartr}
\alias{smartr-package}
\title{Tools for Nutritional Status and Mortality Rate Assessments and Surveys}
\description{
SMART is an inter-agency initiative launched in 2002 by a network of
organizations and humanitarian practitioners. SMART advocates a
multi-partner, systematized approach to provide critical, reliable
information for decision-making, and to establish shared systems and
resources for host government partners and humanitarian organizations. The
SMART Methodology is an improved survey method that balances simplicity
(for rapid assessment of acute emergencies) and technical soundness. This
package provides functions to perform the various analysis used in SMART.
}
\author{
\strong{Maintainer}: Ernest Guevarra \email{ernest@guevarra.io} (\href{https://orcid.org/0000-0002-4887-4415}{ORCID})

}
\keyword{internal}
