% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_gui.R
\name{smartRgui}
\alias{smartRgui}
\title{smartR GUI}
\usage{
smartRgui(smartRstudy = NULL)
}
\arguments{
\item{smartRstudy}{SmartProject class}
}
\value{
This function does not return a value.
}
\description{
The \code{smartRgui} function implements the main graphical user interface
of smartR. Please, contact the package maintainer for detailed instructions.
}
\examples{
\dontrun{
yourSmartRstudy <- SmartProject$new()
smartRgui(smartRstudy = yourSmartRstudy)
}
}
