% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_class.R
\docType{data}
\name{SurveyBySpecie}
\alias{SurveyBySpecie}
\title{SurveyBySpecie}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SurveyBySpecie
}
\value{
This function returns the 'SurveyBySpecie' object.
}
\description{
The \code{SurveyBySpecie} class implements the class of SMART to
 handle species samplings.
}
\keyword{datasets}

