% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/update_job.R
\name{update_job}
\alias{update_job}
\title{Update the job status}
\usage{
update_job(.x, status, path = NULL)
}
\arguments{
\item{.x}{The index or name of the job.}

\item{status}{The status of the job.}

\item{path}{The path to the job log.}
}
\description{
This function will update the status of the job.
}
