% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_manager.R
\name{view_job}
\alias{view_job}
\title{View the status of a job.}
\usage{
view_job(.x = NULL, path = NULL)
}
\arguments{
\item{.x}{Character or numeric. The job name (character) or index (numeric).
If `NULL`, returns all jobs.}

\item{path}{Character. Optional. The directory where the job log is stored.
If `NULL`, the function looks in the temporary directory.}
}
\value{
A data frame containing the status of the specified job(s).
}
\description{
Retrieves job statuses from the job log. If no job index or name is specified, returns all jobs.
}
\examples{
view_job()  # View all jobs
view_job(1)  # View job with index 1
view_job("job_name")  # View job with name "job_name"

}
