% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{villageList1}
\alias{villageList1}
\title{List of villages in Dessie Zuria survey with population sizes and cumulative
population size range}
\format{
A tibble with 281 rows and 3 columns:

| **Variable** | **Description** |
| :--- | :--- |
| *village_id* | Village identifier |
| *village_name* | Name of village |
| *pop* | Village population size |
| *cum_pop* | Cumulative population size range |
}
\source{
SMART, Action Against Hunger-Canada, and Technical Advisory Group
}
\usage{
villageList1
}
\description{
List of villages in Dessie Zuria survey with population sizes and cumulative
population size range
}
\examples{
villageList1


}
\keyword{datasets}
