% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{villageList2}
\alias{villageList2}
\title{Example list of villages with population sizes}
\format{
A tibble with 10 rows and 2 columns:

| **Variable** | **Description** |
| :--- | :--- |
| *village_id* | Village identifier |
| *pop* | Village population size |
}
\source{
SMART, Action Against Hunger-Canada, and Technical Advisory Group
}
\usage{
villageList2
}
\description{
Example list of villages with population sizes
}
\examples{
villageList2

}
\keyword{datasets}
