% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/agg_plot.R
\name{agg_plot}
\alias{agg_plot}
\title{Aggregate data for plotting}
\usage{
agg_plot(data, y, between = NULL, within = NULL, group = NULL, ci = 0.95)
}
\arguments{
\item{data}{A data frame containing the data to be aggregated}

\item{y}{The variable to be aggregated}

\item{between}{The between-subject factor}

\item{within}{The within-subject factor}

\item{group}{The grouping variable}

\item{ci}{The confidence interval}
}
\value{
A data frame containing the aggregated data
}
\description{
This function aggregates data for plotting. It computes the mean, standard error, and confidence interval of the data.
}
