% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job-manager.R
\name{append_job}
\alias{append_job}
\title{Append a new job to the job log.}
\usage{
append_job(
  name = NULL,
  cores = 1,
  untilFinished = NULL,
  priority = 0,
  path = NULL
)
}
\arguments{
\item{name}{Character. The name of the job. This must be provided.}

\item{cores}{Integer. The number of cores allocated to the job. Default is `1`.}

\item{untilFinished}{Character or numeric vector. If it is a numeric vector, it will be converted to a character.}

\item{priority}{Numeric. The priority of the job, with higher numbers indicating higher priority. Default is `0`.}

\item{path}{Character. Optional. The directory where the job log is stored. If `NULL`, the function looks in the temporary directory.}
}
\description{
Adds a new job entry to the job log, assigning an index, priority, and status.
}
\keyword{internal}
