% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_logml.R
\name{compare_logml}
\alias{compare_logml}
\title{Compare log marginal likelihood of different models}
\usage{
compare_logml(pars, sample_path, sample_prefix)
}
\arguments{
\item{pars}{A list of parameters to compare.}

\item{sample_path}{The path to the samples.}

\item{sample_prefix}{The prefix of the samples.}
}
\value{
A table with the log marginal likelihood of the different models.
}
\description{
This function compares the log marginal likelihood of different models.
}
