% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_class.R
\docType{class}
\name{FisheryBySpecie}
\alias{FisheryBySpecie}
\title{FisheryBySpecie}
\format{\code{\link{R6Class}} object.}
\usage{
FisheryBySpecie
}
\value{
Object of \code{\link{R6Class}} with attributes and methods for the fishery data.
}
\description{
The \code{FisheryBySpecie} class implements the class of SMART to
 handle species samplings.
}
\section{Fields}{

\describe{
\item{\code{specie}}{Missing description.}

\item{\code{year}}{Missing description.}

\item{\code{rawLFD}}{Missing description.}

\item{\code{lengClas}}{Missing description.}

\item{\code{LFDPop}}{Missing description.}

\item{\code{mixPar}}{Missing description.}

\item{\code{nCoho}}{Missing description.}

\item{\code{spreDist}}{Missing description.}

\item{\code{sprePlot}}{Missing description.}

\item{\code{spreSpat}}{Missing description.}

\item{\code{sampMcmc}}{Missing description.}

\item{\code{groMixout}}{Missing description.}

\item{\code{groPars}}{Missing description.}

\item{\code{Coh_A}}{Missing description.}

\item{\code{Coh_A_Int}}{Missing description.}

\item{\code{LWpar}}{Missing description.}

\item{\code{LWparQ}}{Missing description.}

\item{\code{selPar}}{Missing description.}
}}

\section{Methods}{

\describe{
  \item{Documentation}{For full documentation of each method go to https://github.com/smartR_blablabla/}
  \item{\code{initialize(sing_spe)}}{This method is used }
  \item{\code{setRawData(raw_data)}}{This method is used }
  \item{\code{plotLFD()}}{This method is used }
  \item{\code{setYears()}}{This method is used }
  \item{\code{setSpecie()}}{This method is used }
  \item{\code{setLClass()}}{This method is used }
  \item{\code{setNCoho(num_coh)}}{This method is used }
  \item{\code{setLWpar(alphaVal, betaVal, sex)}}{This method is used }
  \item{\code{setWeight(sexVal = "Female")}}{This method is used }
  \item{\code{setSpreDistSing()}}{This method is used }
  \item{\code{setSprePlot(sampSex)}}{This method is used }
  \item{\code{setSpatDistSing()}}{This method is used }
  \item{\code{setSpatPlot(sampSex)}}{This method is used }
  \item{\code{setLWstat(lwUnit)}}{This method is used }
  \item{\code{getMCsamps(numSamp, numAdap, numIter, sexDrop, curveSel)}}{This method is used }
  \item{\code{getGrowPar(sexDrop)}}{This method is used }
  \item{\code{getMCage(sexDrop)}}{This method is used }
  \item{\code{setMCplot(sexDrop, selCurve)}}{This method is used }
  \item{\code{calcMixDate(nAdap, nSamp, nIter, sexDrop, curveSel)}}{This method is used }
  \item{\code{ggplotMcmcOut(selCompo, selSex)}}{This method is used }
  }
}

\keyword{data}
