% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_class.R
\docType{class}
\name{FishFleet}
\alias{FishFleet}
\title{FishFleet}
\format{\code{\link{R6Class}} object.}
\value{
Object of \code{\link{R6Class}} with attributes and methods for the fishery data.
}
\description{
The \code{FishFleet} class implements the class of SMART
to manage fleet data.
}
\section{Fields}{

\describe{
\item{\code{rawRegister}}{Missing description.}

\item{\code{vmsRegister}}{Missing description.}

\item{\code{rawEffort}}{Missing description.}

\item{\code{weekEffoMatr}}{Missing description.}

\item{\code{dayEffoMatr}}{Missing description.}

\item{\code{prodMatr}}{Missing description.}

\item{\code{effoProd}}{Missing description.}

\item{\code{effoProdMont}}{Missing description.}

\item{\code{effoMont}}{Missing description.}

\item{\code{effoProdAll}}{Missing description.}

\item{\code{effoAll}}{Missing description.}

\item{\code{trackHarbs}}{Missing description.}

\item{\code{regHarbsUni}}{Missing description.}

\item{\code{regHarbsBox}}{Missing description.}

\item{\code{rawSelectivity}}{Missing description.}

\item{\code{rawProduction}}{Missing description.}

\item{\code{rawEconomy}}{Missing description.}

\item{\code{registerIds}}{Missing description.}

\item{\code{predProd}}{Missing description.}

\item{\code{productionIds}}{Missing description.}

\item{\code{prodIdsLoa}}{Missing description.}

\item{\code{prodSpec}}{Missing description.}

\item{\code{specSett}}{Missing description.}

\item{\code{specLogit}}{Missing description.}

\item{\code{effortIds}}{Missing description.}

\item{\code{idsEffoProd}}{Missing description.}

\item{\code{effoProdAllLoa}}{Missing description.}

\item{\code{effoAllLoa}}{Missing description.}

\item{\code{effortIndex}}{Missing description.}

\item{\code{daysAtSea}}{Missing description.}

\item{\code{prodIndex}}{Missing description.}

\item{\code{resNNLS}}{Missing description.}

\item{\code{betaAvg}}{Missing description.}

\item{\code{effortAvg}}{Missing description.}

\item{\code{betaMeltYear}}{Missing description.}

\item{\code{prodMeltYear}}{Missing description.}

\item{\code{fishPoinPara}}{Missing description.}

\item{\code{ecoPrice}}{Missing description.}

\item{\code{inSpatialReg}}{Missing description.}

\item{\code{inEffortReg}}{Missing description.}

\item{\code{inProductionReg}}{Missing description.}

\item{\code{outSpatialReg}}{Missing description.}

\item{\code{outEffortReg}}{Missing description.}

\item{\code{outProductionReg}}{Missing description.}

\item{\code{plotSpatialReg}}{Missing description.}

\item{\code{plotEffortReg}}{Missing description.}

\item{\code{plotProductionReg}}{Missing description.}
}}

\section{Methods}{

\describe{
  \item{Documentation}{For full documentation of each method go to https://github.com/smartR_blablabla/}
  \item{\code{setVmsRegister()}}{This method is used }
  \item{\code{setRegHarbs()}}{This method is used }
  \item{\code{setEcoPrice(sel_specie, price_df)}}{This method is used }
  \item{\code{saveFleetHarb(harb_path)}}{This method is used }
  \item{\code{loadFleetHarb(harb_path)}}{This method is used }
  \item{\code{setBetaAvg(sel_specie)}}{This method is used }
  \item{\code{setEffortAvg()}}{This method is used }
  \item{\code{loadFleetRegis(register_path)}}{This method is used }
  \item{\code{loadMatEffort(effort_path)}}{This method is used }
  \item{\code{loadRawEconomy(economic_path)}}{This method is used }
  \item{\code{setYearEconomy()}}{This method is used }
  \item{\code{setInSpatial()}}{This method is used }
  \item{\code{setInEffort()}}{This method is used }
  \item{\code{getRegSpatial()}}{This method is used }
  \item{\code{getRegEffort()}}{This method is used }
  \item{\code{getRegProduction()}}{This method is used }
  \item{\code{getCostOutput()}}{This method is used }
  \item{\code{setCostPlot()}}{This method is used }
  \item{\code{loadProduction(production_path)}}{This method is used }
  \item{\code{setFishPoinPara(speed_range, depth_range)}}{This method is used }
  \item{\code{setWeekMonthNum()}}{This method is used }
  \item{\code{setFishPoin()}}{This method is used }
  \item{\code{plotFishPoinStat()}}{This method is used }
  \item{\code{plotSpeedDepth(which_year, speed_range, depth_range)}}{This method is used }
  \item{\code{setEffortIds()}}{This method is used }
  \item{\code{setProdSpec()}}{This method is used }
  \item{\code{setSpecSett()}}{This method is used }
  \item{\code{setNNLS()}}{This method is used }
  \item{\code{setBetaMeltYear(specie)}}{This method is used }
  \item{\code{setProdMeltYear(specie)}}{This method is used }
  \item{\code{plotTotProd(specie)}}{This method is used }
  \item{\code{plotNNLS(specie, thresR2)}}{This method is used }
  \item{\code{setSpecSettItm(specie, thresh, brea, max_xlim)}}{This method is used }
  \item{\code{plotLogitROC(selSpecie)}}{This method is used }
  \item{\code{setSpecLogitConf(selSpecie, cutoff)}}{This method is used }
  \item{\code{setLogitTrain(selSpecie, train, cp_val, cv_val)}}{This method is used }
  \item{\code{setLogitTest(selSpecie, test)}}{This method is used }
  \item{\code{setLogitPred(selSpecie, test)}}{This method is used }
  \item{\code{setLogitCut(selSpecie)}}{This method is used }
  \item{\code{setLogitRoc(selSpecie)}}{This method is used }
  \item{\code{setLogitConf(selSpecie, test)}}{This method is used }
  \item{\code{setSpecLogit(selSpecie, selModel, cp, cv)}}{This method is used }
  \item{\code{getMatSpeLand(specie)}}{This method is used }
  \item{\code{setEffoProdAll()}}{This method is used }
  \item{\code{setEffoAll()}}{This method is used }
  \item{\code{setEffoProdAllLoa()}}{This method is used }
  \item{\code{setEffoAllLoa()}}{This method is used }
  \item{\code{setProdIds()}}{This method is used }
  \item{\code{setIdsEffoProd()}}{This method is used }
  \item{\code{plotCountIDsEffoProd()}}{This method is used }
  \item{\code{plotCountIDsEffo()}}{This method is used }
  \item{\code{plotCountIDsProd()}}{This method is used }
  \item{\code{setEffoProdMatr()}}{This method is used }
  \item{\code{setEffoProdMont()}}{This method is used }
  \item{\code{setEffoMont()}}{This method is used }
  \item{\code{setProdMatr()}}{This method is used }
  \item{\code{setDayEffoMatrGround(maxFG)}}{This method is used }
  \item{\code{getLoa4Prod()}}{This method is used }
  \item{\code{plotLoaProd()}}{This method is used }
  \item{\code{readRegisterEU(reg_path)}}{This method is used }
  \item{\code{cleanRegister()}}{This method is used }
  \item{\code{plotRegSum()}}{This method is used }
  \item{\code{plotBarReg(regVar, p_las, title)}}{This method is used }
  \item{\code{plotBoxReg(regVar, title)}}{This method is used }
  \item{\code{setRegIds()}}{This method is used }
  }
}

\keyword{data}
