% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sampling.R
\name{get_si}
\alias{get_si}
\title{Get sampling interval}
\usage{
get_si(pop, m)
}
\arguments{
\item{pop}{Total population}

\item{m}{Number of clusters to be selected}
}
\value{
An integer value for the sampling interval
}
\description{
Get sampling interval
}
\examples{
get_si(clusterList$pop, m = 30)

}
