% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_manager.R
\name{init_job}
\alias{init_job}
\title{Initialize a job log.}
\usage{
init_job(path = NULL)
}
\arguments{
\item{path}{Character. Optional. The directory where the job log should be stored.
If `NULL`, the log is saved in a temporary directory.}
}
\description{
Creates an empty job log as a data frame with model information and stores it as an RDS file.
}
\examples{
file_path <- init_job()  # Saves in temp directory
file_path <- init_job("path/to/directory")  # Saves in a specified directory

}
