% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jatos-helper.R
\name{read_metaData}
\alias{read_metaData}
\title{Read the files information from the metadata file}
\usage{
read_metaData(metaData_path)
}
\arguments{
\item{metaData_path}{A string specifying the directory containing the metadata file.}
}
\value{
A data frame containing study results.
}
\description{
This function reads a metadata JSON file and extracts relevant study information
}
