% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/job_manager.R, R/remove_job.R
\name{remove_job}
\alias{remove_job}
\title{Remove job(s) from the job log.}
\usage{
remove_job(.x, path = NULL)

remove_job(.x, path = NULL)
}
\arguments{
\item{.x}{The index or name of the job to be removed.}

\item{path}{The path to the job log.}
}
\description{
Removes a job (or multiple jobs) from the job log by index or name.

This function will remove job(s) from the job log.
}
\examples{
remove_job(1)  # Remove job with index 1
remove_job("Task_A")  # Remove job with name "Task_A"

}
