% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_class.R
\docType{data}
\name{SampleMap}
\alias{SampleMap}
\title{SampleMap}
\format{An object of class \code{R6ClassGenerator} of length 24.}
\usage{
SampleMap
}
\value{
This function returns the 'sampleMap' object.
}
\description{
The \code{SampleMap} class implements the class of SMART
to control geographical data.
}
\keyword{datasets}

