% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smart_runFun.R
\name{smart_runFun}
\alias{smart_runFun}
\title{Run the function in job with the waiting list}
\usage{
smart_runFun(
  fun,
  args,
  untilFinished = FALSE,
  core = 1,
  maxCore = NULL,
  priority = 1,
  checkInt = 17,
  name = NULL
)
}
\arguments{
\item{fun}{The function to run in parallel.}

\item{args}{The arguments of the function to run in parallel.}

\item{untilFinished}{Logical. If `TRUE`, the function will not run until the previous code is finished.}

\item{core}{The number of cores required to run the function}

\item{maxCore}{The maximum number of cores that can be used to run the function}

\item{priority}{The priority of the function The code with a higher priority will be run first.}

\item{checkInt}{The interval to check the job log.}

\item{name}{The name of the job.}
}
\description{
This function is used to run the code in parallel with the waiting list.
It relies on the `job` package to manage the job log.
}
