% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smartr-package.R
\docType{package}
\name{smartr-package}
\alias{smartr}
\alias{smartr-package}
\title{smartr: Smart Ways to Run Bayesian Models on JATOS Data}
\description{
More about what it does (maybe more than one line) Use four spaces when indenting paragraphs within the Description.
}
\keyword{internal}
